moved logic to server component
